ddl sql for different storage.

+++ packages/watchmen-storage-oracle/meta-scripts/16.5.19/00005-scheduled_task.ddl.sql
@@ -1,2 +1,2 @@
 ALTER TABLE scheduled_task
-    ADD status BIGINT;
+    ADD status NUMBER(1);
